// demos/convex/html/convex/_generated/scheduled_functions.ts
// Generated by @neovex/codegen. Do not edit by hand.
import { makeMutationReference } from "convex/browser";
import type { Doc, Id } from "./dataModel";

type JsonValue = null | boolean | number | string | JsonValue[] | { [key: string]: JsonValue };

export const scheduledFunctions = {
  messages: {
    send: makeMutationReference<{
  "author": string;
  "body": string;
}, string>("messages:send", "public"),
    scheduleSend: makeMutationReference<{
  "author": string;
  "body": string;
  "delayMs": number;
}, string>("messages:scheduleSend", "public"),
    rename: makeMutationReference<{
  "id": Id<"messages">;
  "body": string;
}, string>("messages:rename", "public"),
    remove: makeMutationReference<{
  "id": Id<"messages">;
}, null>("messages:remove", "public")
  }
};

export const internalScheduledFunctions = {
  messages: {
    sendInternal: makeMutationReference<{
  "author": string;
  "body": string;
}, Id<"messages">>("messages:sendInternal", "internal")
  }
};